Implement MongoDB based mod cleaning
